# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Role:
    properties:
      guildID:
        type: string
      ignored:
        type: boolean
      isDefault:
        type: boolean
      reward:
        type: integer
      roleID:
        type: string
      xpBlacklisted:
        type: boolean
    type: object
...
  models.UserModification:
    properties:
      discordID:
        type: string
      email:
        type: string
      oldPassword:
        type: string
      password:
        type: string
    type: object
...
host: cardinal.gyroskan.com
info:
  contact:
    email: gyroskan@gmail.com
    name: API Support
  description: The API to interact with cardinal discord bot database.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Cardinal API
  version: "0.1"
paths:
...
  /guilds/{guildID}/roles:
    get:
      description: Fetch all roles of the guild.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - default: false
        description: reward from this lvl only
        in: query
        name: reward
        type: boolean
      - default: false
        description: ignored roles only
        in: query
        name: ignored
        type: boolean
      - default: false
        description: xpBlacklist roles only
        in: query
        name: xpBlacklist
        type: boolean
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "500":
          description: Server error
      summary: Get Guild roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new role for a guild.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: Role values
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "201":
          description: Created role
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Wrong values
        "403":
          description: Forbidden
        "500":
          description: Server Error
      summary: Create role
      tags:
      - Roles
  /guilds/{guildID}/roles/{roleID}:
    delete:
      consumes:
      - application/json
      description: Delete a guild role
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: role id
        in: path
        name: roleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "206":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Fountd
        "500":
          description: Server Error
      summary: Delete guild role
      tags:
      - Roles
    get:
      description: Fetch the role of the guild.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: role id
        in: path
        name: roleID
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Server error
      summary: Get one Guild role
      tags:
      - Roles
    patch:
      consumes:
      - application/json
      description: Update fields of a guild's role
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: role id
        in: path
        name: roleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: Not Fountd
        "500":
          description: Server Error
      summary: Update role values
      tags:
      - Roles
  /users/:
    get:
      description: Get a list of existing users
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "500":
          description: Server error
      summary: Get Users
      tags:
      - Users
  /users/{username}:
    delete:
      description: Delete definitively the user.
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: NotFound
        "500":
          description: Server error
      summary: Delete User
      tags:
      - Users
    get:
      description: Get a specific user
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: Not found
        "500":
          description: Server error
      summary: Get User
      tags:
      - Users
    patch:
      description: Update specified User fields
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: fields to modify
        in: body
        name: userModif
        required: true
        schema:
          $ref: '#/definitions/models.UserModification'
      responses:
        "200":
          description: OK
        "400":
          description: Invalid values
        "403":
          description: Forbidden
        "404":
          description: NotFound
        "500":
          description: Server error
      summary: Update User
      tags:
      - Users
    post:
      description: Update User access level
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: access_level
        in: query
        name: access_level
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: NotFound
        "500":
          description: Server error
      summary: Update User access level
      tags:
      - Users
  /users/{username}/ban:
    delete:
      description: Update User ban. POST to unbann, DELETE to ban.
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: NotFound
        "500":
          description: Server error
      summary: Ban User
      tags:
      - Users
    post:
      description: Update User ban. POST to unbann, DELETE to ban.
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: NotFound
        "500":
          description: Server error
      summary: Ban User
      tags:
      - Users
...
  /users/me:
    get:
      description: Get the logged in user
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "500":
          description: Server error
      summary: Get Logged in User
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"